Allow more than 64k in multipart

+++ src/mastodon/api/compose.c
@@ -60,7 +60,7 @@ char *api_send_hgr_image(char *filename, char *description, char **err, char x, 
   surl_multipart_send_num_fields(1);
   
   /* Send file */
-  surl_multipart_send_field_desc("file", to_send, "image/hgr");
+  surl_multipart_send_field_desc("file", (uint32)to_send, "image/hgr");
   while ((r = fread(buf, sizeof(char), sizeof(buf), fp)) > 0) {
 send_again:
     surl_multipart_send_field_data(buf, r);
